fix(rt): dedup terrain patch vertices and drop redundant copy in radiomap payload
Two Copilot findings on PR #26:

- build_terrain_patch no longer flattens every selected triangle into a
  fresh 3-vertex block. Each source mesh's selected faces are now
  deduplicated via np.unique(..., return_inverse=True), and the resulting
  per-tile (positions, faces, normals, texcoords) blocks are concatenated
  with index offsets. Shared-edge topology is preserved within each tile,
  so _subdivide_triangles' edge-midpoint cache works again and vertex
  count / memory scale with unique vertices rather than triangle corners
  for typical manifold terrain meshes. Vertices across tile seams are
  intentionally not merged because tile coordinates may not coincide
  exactly. Per-vertex normals come from the source mesh when available;
  a small _vertex_normals_from_faces fallback computes averaged face
  normals otherwise.

- solve_radiomap's values_payload comprehension no longer goes through
  values_db.astype(float). linear_to_db already returns float64, so the
  extra full-size copy was redundant and roughly doubled peak memory
  during serialization for large radio maps. Iterating values_db
  directly keeps the same non-finite-to-None behaviour without the copy.

+++ backend/rt/solve_radiomap.py
@@ -125,7 +125,7 @@ def solve_terrain_radiomap(
     finite_values_db = values_db[np.isfinite(values_db)]
     values_payload = [
         float(value) if np.isfinite(value) else None
-        for value in values_db.astype(float)
+        for value in values_db
     ]
     log_timing("radiomap_finalize", finalize_started_at, cells=patch_meta["cell_count"])
